Init OMAP4 power management earlier
PandaBoards tend to lock up during boot because of an invalid idle
handler. The OMAP4-specific idle handler is installed late during
kernel initialization, so secondary CPUs might call the invalid
default idle handler.

The power management for the OMAP4 is now initialized earlier as part
of the architecture code. This at least softens the race condition,
and allows to boot all previously broken PandaBoard that were tested
with this fix; in AOSP and Linaro environments.

Signed-off-by: Thomas Zimmermann <tdz@users.sourceforge.net>